network_bench_logger/requirements.txt
flask
requests
kafka-python
geopy
geographiclib
numpy
ujson